extract the model fetching logic into a separate hook
